Publish linux packages in own infra
And add some mintlify docs
