+++ .github/workflows/goreleaser.yaml
@@ -8,43 +8,26 @@ on:
 
 permissions:
   contents: write
+  id-token: write
 
 jobs:
   goreleaser:
-    runs-on: ubuntu-24.04
-    env:
-      CLOUDSMITH_API_KEY: ${{ secrets.CLOUDSMITH_API_KEY }}
-      CLOUDSMITH_NAMESPACE: ${{ vars.CLOUDSMITH_NAMESPACE || 'libops' }}
-      CLOUDSMITH_REPOSITORY: ${{ vars.CLOUDSMITH_REPOSITORY || 'sitectl' }}
-      CLOUDSMITH_DEB_TARGETS: ${{ vars.CLOUDSMITH_DEB_TARGETS || 'debian/bookworm ubuntu/noble' }}
-      CLOUDSMITH_RPM_TARGETS: ${{ vars.CLOUDSMITH_RPM_TARGETS || 'fedora/41' }}
-      CLOUDSMITH_ALPINE_TARGETS: ${{ vars.CLOUDSMITH_ALPINE_TARGETS || 'alpine/any-version' }}
-    steps:
-      - name: Checkout
-        uses: actions/checkout@93cb6efe18208431cddfb8368fd83d5badbf9bfd # v5
-        with:
-          fetch-depth: 0
-
-      - name: Set up Go
-        uses: actions/setup-go@4dc6199c7b1a012772edbd06daecab0f50c9053c # v6
-        with:
-          go-version: '>=1.25.8'
-
-      - name: Run GoReleaser
-        uses: goreleaser/goreleaser-action@e435ccd777264be153ace6237001ef4d979d3a7a # v6
-        with:
-          distribution: goreleaser
-          version: latest
-          args: release --clean
-        env:
-          GITHUB_TOKEN: ${{ secrets.HOMEBREW_REPO }}
-
-      - name: Install Cloudsmith CLI
-        if: ${{ env.CLOUDSMITH_API_KEY != '' }}
-        uses: cloudsmith-io/cloudsmith-cli-action@v1
-        with:
-          api-key: ${{ env.CLOUDSMITH_API_KEY }}
-
-      - name: Publish Linux packages to Cloudsmith
-        if: ${{ env.CLOUDSMITH_API_KEY != '' }}
-        run: make publish-cloudsmith
+    uses: libops/terraform-linux-packages/.github/workflows/reusable-goreleaser.yaml@main
+    permissions:
+      contents: write
+      id-token: write
+    secrets: inherit
+    with:
+      go-version: ">=1.25.8"
+      publish-package-repo: true
+      package-name: sitectl
+      package-repo-prefix: sitectl
+      package-repo-label: sitectl
+      package-public-key-name: sitectl-archive-keyring
+      gcp-project: ${{ vars.LIBOPS_PACKAGES_GCLOUD_PROJECT }}
+      workload-identity-provider: ${{ vars.LIBOPS_PACKAGES_GCLOUD_OIDC_POOL }}
+      service-account: ${{ vars.LIBOPS_PACKAGES_GSA }}
+      gcs-bucket: ${{ vars.LIBOPS_PACKAGES_GCS_BUCKET }}
+      aptly-gpg-key-id: ${{ vars.LIBOPS_PACKAGES_APTLY_GPG_KEY_ID }}
+      aptly-gpg-private-key-secret: ${{ vars.LIBOPS_PACKAGES_APTLY_GPG_PRIVATE_KEY_SECRET }}
+      aptly-gpg-passphrase-secret: ${{ vars.LIBOPS_PACKAGES_APTLY_GPG_PASSPHRASE_SECRET }}
