+++ Makefile
@@ -1,6 +1,7 @@
-.PHONY: build deps lint test docker integration-test docs plugins install-plugins publish-cloudsmith
+.PHONY: build deps lint test docker integration-test docs docs-host plugins install-plugins publish-aptly-repo
 
 BINARY_NAME=sitectl
+DOCS_PORT ?= 3000
 
 deps:
 	go get .
@@ -8,6 +9,17 @@ deps:
 
 build: deps
 	go build -o $(BINARY_NAME) .
+
+docs:
+	docker run --rm -it \
+		-p $(DOCS_PORT):$(DOCS_PORT) \
+		-v "$(CURDIR):/work" \
+		-w /work \
+		node:22-bookworm \
+		sh -lc "npx mint dev --port $(DOCS_PORT) --host 0.0.0.0"
+
+docs-host:
+	npx mint dev
 
 lint:
 	go fmt ./...
@@ -23,5 +35,5 @@ lint:
 test: build
 	go test -v -race ./...
 
-publish-cloudsmith:
-	bash ./scripts/publish-cloudsmith.sh
+publish-aptly-repo:
+	bash ./scripts/publish-aptly-repo.sh
