+++ docs.json
@@ -1,0 +1,100 @@
+{
+  "$schema": "https://mintlify.com/docs.json",
+  "theme": "mint",
+  "name": "sitectl",
+  "colors": {
+    "primary": "#16A34A",
+    "light": "#07C983",
+    "dark": "#15803D"
+  },
+  "favicon": "/favicon.ico",
+  "navigation": {
+    "tabs": [
+      {
+        "tab": "Home",
+        "groups": [
+          {
+            "group": "Getting started",
+            "pages": [
+              "index",
+              "docs/install",
+              "docs/quickstart"
+            ]
+          },
+          {
+            "group": "Concepts",
+            "pages": [
+              "docs/tui",
+              "docs/context",
+              "docs/plugins",
+              "docs/components"
+            ]
+          },
+          {
+            "group": "Usage",
+            "pages": [
+              "docs/commands/compose"
+            ]
+          },
+          {
+            "group": "Plugins",
+            "pages": [
+              "docs/plugins/drupal",
+              "docs/plugins/isle"
+            ]
+          },
+          {
+            "group": "Contributors",
+            "pages": [
+              "docs/contributors",
+              "docs/contributors/docs"
+            ]
+          }
+        ]
+      },
+      {
+        "tab": "Drupal plugin",
+        "groups": [
+          {
+            "group": "Commands",
+            "pages": [
+              "docs/plugins/drupal/drush"
+            ]
+          }
+        ]
+      },
+      {
+        "tab": "ISLE plugin",
+        "groups": [
+          {
+            "group": "Commands",
+            "pages": [
+              "docs/plugins/isle/index"
+            ]
+          },
+          {
+            "group": "Components",
+            "pages": [
+              "docs/plugins/isle/fcrepo",
+              "docs/plugins/isle/blazegraph"
+            ]
+          }
+        ]
+      }
+    ]
+  },
+  "logo": {
+    "light": "/docs/logo/libops-light.png",
+    "dark": "/docs/logo/libops-dark.png"
+  },
+  "contextual": {
+    "options": [
+      "copy"
+    ]
+  },
+  "footer": {
+    "socials": {
+      "github": "https://github.com/libops/sitectl"
+    }
+  }
+}
