+++ docs/components.mdx
@@ -1,0 +1,19 @@
+---
+title: Components
+description: Understand how sitectl uses components to express reviewed stack defaults and operator tradeoffs.
+---
+
+# Components
+
+Components describe optional parts of a stack in a structured way.
+
+They are how `sitectl` marries infrastructure settings and app-specific settings into one reviewed configuration.
+
+Instead of treating a stack as one large blob, `sitectl` lets each component carry its own defaults, follow-up questions, and operator guidance.
+
+Common component states:
+
+- `enabled`
+- `disabled`
+- `superseded`
+- `distributed`
