+++ docs/contributors/docs.mdx
@@ -1,0 +1,61 @@
+---
+title: Documentation
+description: How to edit, preview, and organize the Mintlify documentation in the sitectl repository.
+---
+
+import { TUI } from "/docs/snippets/tui-tooltip.mdx";
+
+The docs site is driven by Mintlify from this repository.
+
+Mintlify's docs are available at https://www.mintlify.com/docs
+
+Mintlify reads:
+
+- [docs.json](https://github.com/libops/sitectl/blob/main/docs.json) for navigation and site settings
+- [index.mdx](https://github.com/libops/sitectl/blob/main/index.mdx) for the homepage
+- files under [`docs/`](https://github.com/libops/sitectl/tree/main/docs) for the rest of the documentation
+
+## Local development
+
+Mintlify’s local dev startup command is `mint dev`, but Mintlify does not support Node 25+. This repo avoids that issue by running Mintlify in a Node 22 Docker container:
+
+```bash
+make docs
+```
+
+The docs server binds to port `3000` by default. You can override it with:
+
+```bash
+make docs DOCS_PORT=3333
+```
+
+If you already have a supported LTS Node version installed locally, you can also run the host-native command:
+
+```bash
+make docs-host
+```
+
+## Making docs changes
+
+When you update the docs:
+
+- keep pages in `docs/` unless the page is the root homepage, which lives at `index.mdx`
+- preserve the existing nav structure in `docs.json` and only add to it when needed
+- use MDX for new pages
+- keep contributor and operator guidance in the docs site instead of reviving `CONTRIBUTING.md`
+
+## <TUI /> content
+
+The <TUI /> tour reads:
+
+- `internal/tuitour/content/index.md`
+- `internal/tuitour/content/tui.md`
+- `internal/tuitour/content/context.md`
+- `internal/tuitour/content/plugins.md`
+- `internal/tuitour/content/components.md`
+
+That content is intentionally separate from the Mintlify docs so the in-app tour can stay plain markdown and keep its own presentation.
+
+## Deployment
+
+Docs deploy through the Mintlify GitHub App connected to this repository. There is no GitHub Pages workflow to maintain here.
