+++ docs/install.mdx
@@ -1,0 +1,69 @@
+---
+title: Install
+description: Install sitectl with Homebrew, native Linux packages, or direct binaries.
+---
+
+## Homebrew
+
+You can install `sitectl` using Homebrew:
+
+```bash
+brew tap libops/homebrew https://github.com/libops/homebrew
+brew install libops/homebrew/sitectl
+```
+
+## Linux Packages
+
+Releases publish native Linux packages through the libops package repository.
+
+<CodeGroup>
+
+```bash Debian / Ubuntu
+curl -fsSL https://packages.libops.io/sitectl/sitectl-archive-keyring.asc | sudo gpg --dearmor -o /usr/share/keyrings/sitectl-archive-keyring.gpg
+echo "deb [signed-by=/usr/share/keyrings/sitectl-archive-keyring.gpg] https://packages.libops.io/sitectl ./" | sudo tee /etc/apt/sources.list.d/sitectl.list >/dev/null
+sudo apt update
+sudo apt install sitectl
+```
+
+```bash Fedora / Rocky / RHEL
+sudo tee /etc/yum.repos.d/sitectl.repo >/dev/null <<'EOF'
+[sitectl]
+name=sitectl
+baseurl=https://packages.libops.io/sitectl/rpm
+enabled=1
+gpgcheck=0
+repo_gpgcheck=1
+gpgkey=https://packages.libops.io/sitectl/sitectl-archive-keyring.asc
+EOF
+
+sudo dnf makecache
+sudo dnf install sitectl
+```
+
+</CodeGroup>
+
+## Binary Install
+
+You can install sitectl by either downloading or building the `sitectl` binary
+
+<AccordionGroup>
+  <Accordion title="Download" icon="download">
+
+You can download a binary for your system from [the latest release of sitectl](https://github.com/libops/sitectl/releases/latest).
+  </Accordion>
+
+  <Accordion title="Build" icon="screwdriver-wrench">
+
+Requires `go` and `make`
+
+```bash
+git clone https://github.com/libops/sitectl
+cd sitectl
+make build
+./sitectl --help
+```
+</Accordion>
+</AccordionGroup>
+
+
+Once `sitectl` is on your system the you can put the binary in a directory that is in your `$PATH`.
