+++ docs/plugins.mdx
@@ -1,0 +1,13 @@
+---
+title: Plugins
+description: How sitectl plugins extend the core CLI with stack-specific workflows.
+---
+
+Plugins extend `sitectl` with stack-specific commands and create flows.
+
+Examples:
+
+- `drupal` adds Drupal-oriented utilities
+- `isle` adds Islandora and ISLE workflows
+
+The core binary discovers installed plugins and exposes them as `sitectl <plugin> ...`.
