+++ docs/plugins/drupal.mdx
@@ -1,0 +1,14 @@
+---
+title: Drupal plugin
+description: Drupal-oriented sitectl workflows and command surface.
+---
+
+# Drupal plugin
+
+The Drupal plugin extends `sitectl` with Drupal-oriented workflows and commands.
+
+Use it when you want Drupal-specific operations layered on top of the core context and compose model.
+
+See also:
+
+- [Drush command](/docs/plugins/drupal/drush)
