+++ docs/plugins/drupal/drush.mdx
@@ -1,0 +1,10 @@
+---
+title: Drush
+description: Drupal plugin command reference entry point for Drush-oriented workflows.
+---
+
+# Drush
+
+The Drupal plugin exposes Drush-oriented workflows through `sitectl`.
+
+This page is the landing page for Drupal command documentation and can be expanded as the plugin surface grows.
