+++ docs/plugins/isle.mdx
@@ -1,0 +1,14 @@
+---
+title: ISLE plugin
+description: Islandora and ISLE-oriented sitectl workflows and component guides.
+---
+
+# ISLE plugin
+
+The ISLE plugin extends `sitectl` with Islandora and ISLE-specific workflows.
+
+See also:
+
+- [ISLE overview](/docs/plugins/isle/index)
+- [Fcrepo component](/docs/plugins/isle/fcrepo)
+- [Blazegraph component](/docs/plugins/isle/blazegraph)
