+++ docs/plugins/isle/blazegraph.mdx
@@ -1,0 +1,10 @@
+---
+title: Blazegraph
+description: Reference anchor for the Blazegraph component in ISLE-oriented sitectl stacks.
+---
+
+# Blazegraph
+
+Blazegraph is another core ISLE service concept.
+
+This page exists as the anchor for ISLE component documentation and can be expanded with deployment and operator guidance.
