+++ docs/plugins/isle/fcrepo.mdx
@@ -1,0 +1,10 @@
+---
+title: Fcrepo
+description: Reference anchor for the Fcrepo component in ISLE-oriented sitectl stacks.
+---
+
+# Fcrepo
+
+Fcrepo is one of the core service concepts that appears in ISLE-oriented stacks.
+
+This page exists as the anchor for ISLE component documentation and can be expanded with stack-specific operator guidance.
