+++ docs/plugins/isle/index.mdx
@@ -1,0 +1,10 @@
+---
+title: ISLE plugin overview
+description: Overview of the sitectl ISLE plugin and where it fits in Islandora-oriented stacks.
+---
+
+# ISLE plugin overview
+
+The ISLE plugin adds Islandora-specific behavior on top of the core `sitectl` workflow.
+
+It is intended for sites and stacks that need Islandora-aware create flows and component guidance.
