+++ docs/quickstart.mdx
@@ -1,0 +1,68 @@
+---
+title: Quickstart
+description: Learn the sitectl command shape and the core areas to explore after installation.
+---
+
+import { TUI } from "/docs/snippets/tui-tooltip.mdx";
+
+## Setup
+
+After [installing `sitectl`](/docs/install), start by running:
+
+```bash
+sitectl
+```
+
+Running `sitectl` with no arguments opens the <TUI />, where you can:
+
+<Columns cols={2}>
+  <Card title="Set up an existing site" icon="server">
+    Connect sitectl to a Compose project you already run locally or remotely.
+  </Card>
+  <Card title="Create a new site" icon="square-plus">
+    Use the guided flow to start a new site with the available stack-specific tooling.
+  </Card>
+</Columns>
+
+## Monitoring
+
+After you setup or install a site, you can use sitectl to monitor the health of the docker compose project.
+
+Information like general container health, host stats, and logs will all be available in the sitectl dashboard.
+
+## Operations
+
+After you've configured your site(s) and environment(s) as sitectl contexts, you can begin using sitectl to help with site operations.
+
+You can also run individual commands directly for one-off tasks when you do not need the full <TUI /> flow.
+
+The main areas to learn from there are:
+
+<Columns cols={2}>
+  <Card title="Contexts" icon="server">
+    Describe a site and its environments so sitectl knows how to connect and operate.
+  </Card>
+  <Card title="Compose commands" icon="docker">
+    Run Docker Compose commands through sitectl against local and remote contexts.
+  </Card>
+  <Card title="Plugins" icon="plug">
+    Add stack-specific behavior for the technologies your sites use.
+  </Card>
+  <Card title="Components" icon="database">
+    Work with reviewed operator defaults for optional parts of a stack.
+  </Card>
+</Columns>
+
+## Basic command shape
+
+```bash
+sitectl <command> [subcommand] [flags]
+```
+
+Common top-level commands:
+
+- `compose`
+- `config`
+- `make`
+- `port-forward`
+- `sequelace`
