+++ docs/snippets/tui-tooltip.mdx
@@ -1,0 +1,8 @@
+export const TUI = () => (
+  <Tooltip
+    headline="TUI"
+    tip="Terminal User Interface: an interactive command-line interface for navigating and operating sitectl."
+  >
+    TUI
+  </Tooltip>
+);
