+++ docs/tui.mdx
@@ -1,0 +1,15 @@
+---
+title: Terminal UI
+description: The interactive terminal interface for moving between sites, environments, and operator actions.
+---
+
+import { TUI } from "/docs/snippets/tui-tooltip.mdx";
+
+The `sitectl` <TUI /> is the main interface for managing a site and its environments.
+
+It gives you one place to:
+
+- inspect environments
+- run commands against a specific environment
+- install new sites
+- move between local and remote environments for the same site
