+++ index.mdx
@@ -1,0 +1,53 @@
+---
+title: sitectl
+description: Command line utility to interact with your local and remote Docker Compose sites.
+---
+
+import { TUI } from "/docs/snippets/tui-tooltip.mdx";
+
+## Overview
+
+Many open source applications can run well on Docker Compose.
+
+Compose also offers a strong developer experience because the same workload can often run in development and production with only modest environment-specific differences.
+
+## The gap
+
+The friction starts when a team is running several Compose projects across local and remote environments.
+
+There is usually enough toil around contexts, repeated operational commands, and stack-specific behavior to justify better tooling, but not enough complexity to justify Kubernetes.
+
+## Why sitectl exists
+
+`sitectl` is a tool for technoligists running a handful of sites using docker compose. Using sitectl is a value add to Docker Compose; the Compose-first workflow remains intact while making it easier to operate that stack day to day.
+
+<Columns cols={2}>
+  <Card title="Interactive operations" icon="terminal">
+    Use the <TUI /> for routine site setup, monitoring, and operator workflows.
+  </Card>
+  <Card title="Contexts" icon="server">
+    Track local and remote environments so sitectl can understand where a site lives and how to reach it.
+  </Card>
+  <Card title="Plugins" icon="plug">
+    Add stack-specific behavior for common technologies without abandoning the core Compose workflow.
+  </Card>
+  <Card title="Components" icon="database">
+    Model reviewed stack defaults and operator choices in a more structured way than ad hoc Compose notes.
+  </Card>
+</Columns>
+
+## Why sitectl vs Docker Context?
+
+While [Docker's native context feature](https://docs.docker.com/engine/manage-resources/contexts/) handles basic docker daemon connections, `sitectl` is purpose-built for Docker Compose projects and adds:
+
+<Columns cols={3}>
+  <Card title="Remote operations" icon="ethernet">
+    SFTP file operations, sudo support, and clearer SSH error handling.
+  </Card>
+  <Card title="Container utilities" icon="toolbox">
+    General helpers to do things like resolve service names to containers, extract secrets and env vars for `exec` commands, and inspect container network details.
+  </Card>
+  <Card title="Compose-first design" icon="code">
+    Automatically set the equivalent of `DOCKER_HOST`, `COMPOSE_PROJECT_NAME`, `COMPOSE_FILE`, and `COMPOSE_ENV_FILES` from the active sitectl context.
+  </Card>
+</Columns>
