+++ internal/tuitour/content/components.md
@@ -12,12 +12,3 @@ Common component states are:
 - `disabled`: the component is not used
 - `superseded`: another component replaces its role
 - `distributed`: responsibility is moved out to external or split services
-
-When you change a default component, you should expect operator implications such as:
-
-- different infrastructure requirements
-- different app-level wiring or environment values
-- data movement or migration work
-- different maintenance and failure modes
-
-That is why component review matters: changing one default can affect both platform behavior and application behavior at the same time.
